chore: adjust script for waiting db & initializing db
- refactor initdb.sh to use Taskfile envs for list of DBs

The idea of this changes is to use available envs from `Taskfile`, then
propagate it to `compose.yaml` as envs (evaluated dynamically when
running `task`). This changes doesn't hardcode sql script but rather use
the env to populate correct db name as before.

The impact of this changes is we need to run `db` from `task` at least
initially since it relied on taskfile specific envs, however for
consecutive runs it's ok to run not from task, since it will fallback to
`pg_isready` if either envs not exists.

- introduce new script named `checkdb.sh`.

The idea of this script is to check whether db is already populated or not,
so rather than using `pg_isready`, we just directly checks whether the
database is populated with correct db or not.